+++ CHANGELOG.md
@@ -1,9 +1,19 @@
-Version: 1.0.1
+# CHANGELOG
+
+### Version: 1.1.1
+Date: 02.04.2022
+New:
+1) Realize message reading
+2) Render time when messages were sent
+3) Scroll messages when chat is opened
+4) Scroll when new message is sent
+
+### Version: 1.0.1
 Date: 29.03.2022
 New:
 1) Change some style
 
-Version: 1.0.0
+### Version: 1.0.0
 Date: 28.03.2022
 New: 
 1) Basic messenger